+++ sample-3D/src/PongGame/Pong.ts
