+++ sample-3D/src/PongGame/PongUpdater.ts
